+++ NEWS.md
@@ -1,4 +1,4 @@
-# workflowsets (development version)
+# workflowsets 1.0.0
 
 * New `extract_parameter_set_dials()` and `extract_parameter_dials()` methods to 
   extract parameter sets and single parameters from `workflow_set` objects.